Python_projects/skud_bot/requirements.txt
Flask==2.2.2
python-dotenv==0.21.0
mysql-connector-python==8.0.30
python-telegram-bot==20.0
requests==2.28.1